+++ packages/delivery-service/src/ws/WebSocketManager.ts
@@ -1,4 +1,4 @@
-import { checkToken } from '@dm3-org/dm3-lib-delivery';
+import { checkToken } from '@dm3-org/dm3-lib-server-side';
 import { normalizeEnsName } from '@dm3-org/dm3-lib-profile';
 import { IWebSocketManager } from '@dm3-org/dm3-lib-shared';
 import { ethers } from 'ethers';
