+++ packages/lib/delivery/src/index.ts
@@ -9,7 +9,6 @@ export { getConversationId } from './Messages';
 export type {} from './PublicMessages';
 export * as schema from './schema';
 export * as spamFilter from './spam-filter/';
-export { checkToken } from './Session';
 export type { Session } from './Session';
 export type { DeliveryServiceProperties } from './Delivery';
 export * from './notifications';
