+++ packages/lib/server-side/package.json
@@ -12,7 +12,6 @@
     }
   },
   "dependencies": {
-    "@dm3-org/dm3-lib-delivery": "workspace:^",
     "@dm3-org/dm3-lib-profile": "workspace:^",
     "@dm3-org/dm3-lib-shared": "workspace:^",
     "express": "^4.18.1"
