+++ packages/lib/server-side/src/index.ts
@@ -1,5 +1,5 @@
 export { Auth } from './auth';
-export { authorize } from './authorize';
+export * from './authorize';
 export { getCachedWebProvider } from './web3Provider/getCachedWebProvider';
 export type { IAccountDatabase } from './iAccountDatabase';
 export * from './utils';
